requirements.txt
fastapi
uvicorn
pandas
gspread
openai
twilio
apscheduler
python-dotenv
pydantic-settings
